Data_Types

Description

Menthors : prof. dr. Andrej Bauer and asist. dr. Matija Pretnar

Authors : Petra Poklukar, Jure Kukovec

Definitons of fundamental datatypes used in SearchMonad, Find_X and Functions.

Documentation

data Bit

Constructors

Zero

One

type Natural = Integer

type Cantor = Natural -> Bit

(#) :: Bit -> Cantor -> Cantor

Returns the representation of the Bit sequence obtained by prepending x to the Bit sequence represented by a.

data T x

Leafless binary tree data type. Each tree has a node and two children, who are again trees.

Constructors

B x (T x) (T x)

code :: (Natural -> x) -> T x

Sequence-to-tree isomorphism.

decode :: T x -> Natural -> x

Inverse of code.

id' :: (Natural -> x) -> Natural -> x

The identity composition of decode and code.