Find_X

Description

Menthors: prof. dr. Andrej Bauer and asist. dr. Matija Pretnar

Authors: Petra Poklukar, Jure Kukovec

Various improved versions of find_i defined in Functions.

Various implemented search functions, improving on find_i defined in Functions.

find_ii :: (Cantor -> Bool) -> Cantor

A mild improvement over find_i, with no calls to exists.

find_iii :: (Cantor -> Bool) -> Cantor

Saves the result yielded by the evaluation of the given predicate, to be used when needed.

find_iv :: (Cantor -> Bool) -> Cantor

An improvement over find_i, but worse that find_iii, due to not having pruned its branching via laziness.

find_v :: (Cantor -> Bool) -> Cantor

Calculates the minimal number of other positions required, to yield the desired one. Starting from the requested position it calculates prior positions as needed and constructs the tail of the solution recursively.

find_vii :: (Cantor -> Bool) -> Cantor

Improved find_v with memoisation using the T data type.