Functions

Description

Menthors: prof. dr. Andrej Bauer and asist. dr. Matija Pretnar

Authors: Petra Poklukar, Jure Kukovec

Basic functions for searching Cantor spaces and auxiliary functions for efficiency testing.

Functions find, forall (universal quantification) and exists (exsistential quantification)

exists :: (Cantor -> Bool) -> Bool

The function exists evaluates to One if a binary sequence satisfying the predicate exists and to Zero otherwise.

forall :: (Cantor -> Bool) -> Bool

The function forall evaluates to One if every binary sequence satisfies the predicate and to Zero otherwise.

find :: (Cantor -> Bool) -> Cantor

Given that exists evaluates to One, find evaluates to some element satisfying the predicate. Even if it evaluates to Zero, find will always return an element of the Cantor set. Therefore, elements yielded by find must additionally be tested whether or not they satisfy the predicate.

find_i :: (Cantor -> Bool) -> Cantor

The function find_i determines that if a solution beginning with Zero exists, the result must also begin with Zero. Otherwise, it must begin with One.

Functions used for testing

search :: (Cantor -> Bool) -> Maybe Cantor

The function search expands the functionality of find and returns a Just value if the result of find, which always yields a value in the Cantor space, satisfies the predicate and Nothing otherwise.

equal :: Eq y => (Cantor -> y) -> (Cantor -> y) -> Bool

Tests function equality. Two functions f and g are equal iff the values f(x) and g(x) are equal for all x in the domain.

implies :: Bool -> Bool -> Bool

The definition of implication.

eq :: Natural -> Cantor -> Cantor -> Bool

The value of eq n a b evaluates to True if the sequences a and b match up until the n-th term and to False otherwise.

least :: (Natural -> Bool) -> Natural

Given a Boolean sequence p, the function least returns the index of the first term of p that equals True.

modulus :: (Cantor -> Integer) -> Natural

Assuming f is an Integer-valued function defined on the Cantor space, modulus f returns the smallest natural number n, such that for any two elements of the Cantor space a and b, a i == b i for all i <= n implies f(a) = f(b).

Alternative implementations of the functions exists and find given that a search function were to be implemented independently.

exists_search :: (Cantor -> Bool) -> Bool

The function takes the result produced by the now predefined search and maps any Just value to True and Nothing to False.

find_search :: (Cantor -> Bool) -> Cantor

The function find_search is an analogy to find_i, where exists_search replaces exists.