SearchMonad

Description

Menthors: prof. dr. Andrej Bauer and asist. dr. Matija Pretnar

Authors: Petra Poklukar, Jure Kukovec

Contains the monad for arbitrary searchable spaces.

Documentation

newtype S a

S over some type a (e.g. Cantor) is a record with exactly one field, which represents its find function (e.g. find_i defined in Find_X for Cantor spaces).

Constructors

S

Fields

find :: (a -> Bool) -> a

exists_mon :: S a -> (a -> Bool) -> Bool

Existential and universal quantifiers defined for searchable sets.

forall_mon :: S a -> (a -> Bool) -> Bool

Existential and universal quantifiers defined for searchable sets.

singleton :: a -> S a

Cannonical mapping of an element into a searchable set. Any predicate tested on a singelton {x} must always return x since it must return something even if the predicate is not satisfied.

search :: S a -> (a -> Bool) -> Maybe a

Modified find which returns Nothing if the predicate is not satisfied for any element and a Just value otherwise.

union :: S a -> S a -> S a

Set-theoretical definition of union where bigUnion creates the union of a familiy of sets.

doubleton :: a -> a -> S a

Searching a doubleton {x, y} for an element that satisfies a given predicate p yields x if p(x) holds and must yield some element of {x,y} otherwise, therefore it returns y if p(x) does not hold.

image :: (a -> b) -> S a -> S b

Given that a subset of a is searchable, the f-image is also searchable. One can find an element in the f-image satisfying a given predicate q by searching the original set for elements satisfying a predicate testing whether q holds for f(x) where x is of type a.

bigUnion :: S (S a) -> S a

Since none of the functions defined above are restricted to the Cantor set we implement the search over an abstract union in two stages. First, using an auxiliary predicate, we find a member of a given family of sets, that contains some element satisfying the original predicate (note that this gives us no answer as to which element it is at this point). Then, this set can again be searched for the particular element.